Changes from the README.md
I've updated the `cook help usage` and please check the changes on windows.

Changed some orders from easier to hard. 

The METHODS are added and the gob date from `(` to `[` in the doc.

Is it possible to use an url in cook and if so, how to use it? 

I'm using this at this moment:
```bash
$ curl -Ss https://raw.githubusercontent.com/danielmiessler/SecLists/master/Fuzzing/User-Agents/operating-platform/zte-z740.txt | cook -param1 \# -param3 testingtestingtesting param1 -d - d param3
#Mozilla/5.0 (Linux; U; Android 4.1.2; en-us; ZTE-Z740 Build/JZO54K) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30testingtestingtesting
#Mozilla/5.0 (Linux; U; Android 4.1.2; en-us; ZTE-Z740 Build/JZO54K) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30 Chrome/33.0.0.0testingtestingtesting
#Mozilla/5.0 (Linux; U; Android 4.1.2; es-us; ZTE-Z740 Build/JZO54K) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30testingtestingtesting
```

Anyways thanks for the awesome framework, i'm trying to use it more and more!